refactor(bridge): extract ethercat_endpoint, runtime_caps, state, motion_caps submodules
Pure code motion out of bridge.rs (4472 -> 3783 lines) plus a with_mcu
helper replacing 11 copies of the lock-map-lookup-error boilerplate.
Error messages unchanged.
