+++ rust/motion-engine/src/bridge.rs
@@ -22,16 +22,27 @@ use crate::mcu_config::{McuAxisConfig, McuCaps, build_mcu_configs};
 use crate::stream_worker::{StreamWorkerError, StreamWorkerHandle};
 use crate::types::{cq_id_from_raw, mcu_handle_from_raw, stats_to_pydict};
 
-struct HomingRun {
-    cohort: u64,
-    endstop_id: u8,
-    endstop_mcu: u32,
-    axis: u8,
-    axis_key: crate::types::AxisKey,
-    all_axis_keys: Vec<crate::types::AxisKey>,
-    window_start_host: f64,
-    notify: crossbeam_channel::Sender<Result<([f64; 3], [f64; 3], u64), String>>,
-}
+mod ethercat_endpoint;
+mod motion_caps;
+mod runtime_caps;
+mod state;
+
+use ethercat_endpoint::{
+    EndpointClaimError, arm_endpoint_death_watchdog, endpoint_args, handshake_ethercat_endpoint,
+    message_for_claim_error, poll_socket_ready, report_ethercat_endpoint_death,
+    spawn_ethercat_endpoint,
+};
+use motion_caps::{
+    axis_ring_depth, drip_cohort_participants, require_events_dir_for_mcu_transport,
+    resolve_motion_caps, ring_depth_for_axis_inner,
+};
+use runtime_caps::{
+    collect_motor_positions_inner, place_motor_response, query_ethercat_runtime_caps,
+    query_runtime_caps, require_positive, slot_for_axis,
+};
+use state::{
+    EngineEvent, EthercatDrive, FlushWait, HomingRun, McuConnection, trip_position_to_motor_frame,
+};
 
 fn abort_after_tracing_appender_drains() {
     let _ = std::io::Write::flush(&mut std::io::stderr());
@@ -41,543 +52,9 @@ fn abort_after_tracing_appender_drains() {
     }
 }
 
-/// How long the host waits for klippy to consume the latched endpoint-death
-/// cause (clean shutdown) before the watchdog forces a last-resort abort. Sized
-/// well above the `DRIVE_FAULT_POLL_PERIOD` (1 s) so a healthy reactor always
-/// shuts down cleanly first; the abort only fires if the reactor is wedged.
-const ENDPOINT_DEATH_SHUTDOWN_GRACE: Duration = Duration::from_secs(5);
-
-/// Latch an EtherCAT-endpoint-death cause for klippy to surface as the shutdown
-/// reason (first cause wins), and log it. Deliberately does NOT abort here: the
-/// host shuts down cleanly via `ethercat_node._poll_drive_fault` →
-/// `invoke_shutdown` so the operator sees the real cause and runs
-/// `FIRMWARE_RESTART` (no silent auto-restart). Returns `true` when this call
-/// latched the first cause, so the caller arms the safety watchdog exactly once.
-fn report_ethercat_endpoint_death(
-    latch: &Arc<Mutex<HashMap<u32, String>>>,
-    mcu_id: u32,
-    reason: &str,
-) -> bool {
-    let code = runtime::error::FaultCode::EthercatEndpointDied.as_i32();
-    let message = format!("EtherCAT endpoint died mid-session (fault {code}): {reason}");
-    let mut guard = latch.lock().unwrap_or_else(|p| p.into_inner());
-    // First cause wins for BOTH the latched (operator-surfaced) message and the
-    // log: a later writer (e.g. the supervisor after the pump already latched)
-    // must not overwrite the original cause.
-    if let std::collections::hash_map::Entry::Vacant(slot) = guard.entry(mcu_id) {
-        slot.insert(message);
-        tracing::error!(
-            subsystem = "ethercat",
-            event = "endpoint_death",
-            mcu_id,
-            fault_code = code,
-            reason,
-            "EtherCAT endpoint died mid-session — latched for klippy; clean shutdown, no abort"
-        );
-        true
-    } else {
-        false
-    }
-}
-
-/// Safety backstop for the clean-shutdown path: if klippy has not consumed the
-/// latched endpoint-death cause within the grace (i.e. the reactor never ran the
-/// shutdown — wedged/CPU-starved), force a last-resort abort so the machine still
-/// stops. The normal path consumes the latch within one poll period, so this only
-/// fires on a double failure (endpoint dead AND reactor stuck).
-fn arm_endpoint_death_watchdog(latch: Arc<Mutex<HashMap<u32, String>>>, mcu_id: u32) {
-    let _ = std::thread::Builder::new()
-        .name(format!("ec-death-watchdog-{mcu_id}"))
-        .spawn(move || {
-            std::thread::sleep(ENDPOINT_DEATH_SHUTDOWN_GRACE);
-            let unhandled = latch
-                .lock()
-                .unwrap_or_else(|p| p.into_inner())
-                .contains_key(&mcu_id);
-            if unhandled {
-                tracing::error!(
-                    subsystem = "ethercat",
-                    event = "endpoint_death_watchdog_abort",
-                    mcu_id,
-                    grace_secs = ENDPOINT_DEATH_SHUTDOWN_GRACE.as_secs(),
-                    "klippy did not act on the latched EtherCAT endpoint death within the grace \
-                     — aborting as a last-resort safety stop"
-                );
-                abort_after_tracing_appender_drains();
-            }
-        });
-}
-
-fn trip_position_to_motor_frame(
-    axis: u8,
-    motor_pos: f64,
-    _configs: &[crate::mcu_config::McuAxisConfig],
-    _axis_mcu: u32,
-) -> [f64; SPATIAL_AXES] {
-    assert!(
-        (axis as usize) < SPATIAL_AXES,
-        "follower axis {axis} in homing trip is a bug — a follower axis must never reach homing recovery"
-    );
-    let mut frame = [0.0f64; SPATIAL_AXES];
-    frame[axis as usize] = motor_pos;
-    frame
-}
-
-struct McuConnection {
-    label: String,
-    serial_path: String,
-    baud: u32,
-    host_io: Option<Arc<McuHostIo>>,
-    runtime_rx_priority: Option<Receiver<host_rt::host_io::runtime_events::RuntimeEvent>>,
-    runtime_rx_bulk: Option<Receiver<host_rt::host_io::runtime_events::RuntimeEvent>>,
-    runtime_caps: Option<mcu_protocol::messages::RuntimeCapsResponse>,
-    identify_caps: u64,
-    mcu_transport_supported: bool,
-    ethercat_socket: Option<String>,
-    endpoint_process: Option<std::process::Child>,
-    endpoint_conn: Option<Arc<McuSerialConn>>,
-    ethercat_slot_axes: Vec<usize>,
-}
-
 const DRAIN_TIMEOUT: Duration = Duration::from_secs(60);
 
-type EthercatDrive = (
-    i32,
-    usize,
-    f64,
-    f64,
-    Option<u32>,
-    Option<u16>,
-    bool,
-    f64,
-    bool,
-    Option<String>,
-);
-
 const ETHERCAT_CLOCK_FREQ_HZ: u32 = 1_000_000_000;
-
-#[derive(Debug, thiserror::Error)]
-enum RuntimeCapsError {
-    #[error("mcu_call QueryRuntimeCaps: {0}")]
-    Call(String),
-    #[error("QueryRuntimeCaps: unexpected response kind {got:?}")]
-    UnexpectedKind { got: mcu_protocol::MessageKind },
-    #[error("decode RuntimeCapsResponse: {0}")]
-    Decode(String),
-}
-
-fn require_positive(value: Option<f64>, name: &str) -> PyResult<()> {
-    if let Some(v) = value {
-        if !(v.is_finite() && v > 0.0) {
-            return Err(PyValueError::new_err(format!(
-                "{name} must be finite and positive"
-            )));
-        }
-    }
-    Ok(())
-}
-
-fn decode_runtime_caps_body(
-    body: &[u8],
-) -> Result<mcu_protocol::messages::RuntimeCapsResponse, RuntimeCapsError> {
-    use mcu_protocol::codec::{Cursor, Decode};
-    use mcu_protocol::messages::RuntimeCapsResponse;
-    let mut c = Cursor::new(body);
-    RuntimeCapsResponse::decode_from(&mut c).map_err(|e| RuntimeCapsError::Decode(format!("{e:?}")))
-}
-
-fn query_runtime_caps(
-    io: &McuHostIo,
-    timeout: std::time::Duration,
-) -> Result<mcu_protocol::messages::RuntimeCapsResponse, RuntimeCapsError> {
-    use mcu_protocol::MessageKind;
-    let (kind, body) = io
-        .mcu_call(MessageKind::QueryRuntimeCaps, Vec::new(), timeout)
-        .map_err(|e| RuntimeCapsError::Call(format!("{e:?}")))?;
-    if kind != MessageKind::RuntimeCapsResponse {
-        return Err(RuntimeCapsError::UnexpectedKind { got: kind });
-    }
-    decode_runtime_caps_body(&body)
-}
-
-enum MotorQuery {
-    Serial(Arc<McuHostIo>),
-    EtherCat(Arc<McuSerialConn>),
-}
-
-impl MotorQuery {
-    fn is_ethercat(&self) -> bool {
-        matches!(self, MotorQuery::EtherCat(_))
-    }
-}
-
-fn slot_for_axis(slot_axes: &[usize], axis: usize) -> Option<u8> {
-    slot_axes
-        .iter()
-        .position(|&a| a == axis)
-        .and_then(|s| u8::try_from(s).ok())
-}
-
-fn place_motor_response(
-    resp: &mcu_protocol::messages::MotorStateResponse,
-    cfg_axes: &[usize],
-    is_ethercat: bool,
-    motors: &mut [Option<f64>],
-    vmotors: &mut [Option<f64>],
-) {
-    let mut put = |slot: usize, m: &mcu_protocol::messages::MotorSample| {
-        if slot < motors.len() {
-            motors[slot] = Some(f64::from(m.pos_q16) / 65536.0);
-            vmotors[slot] = Some(f64::from(m.vel_q16) / 65536.0);
-        }
-    };
-    if is_ethercat {
-        for m in &resp.motors {
-            if let Some(&axis) = cfg_axes.get(m.slot as usize) {
-                put(axis, m);
-            }
-        }
-    } else {
-        for m in &resp.motors {
-            put(m.slot as usize, m);
-        }
-    }
-}
-
-fn collect_motor_positions_inner(
-    mcu_axis_configs: &Mutex<Vec<crate::mcu_config::McuAxisConfig>>,
-    mcus: &Mutex<HashMap<u32, McuConnection>>,
-    timeout: std::time::Duration,
-) -> Result<HashMap<String, (f64, f64)>, String> {
-    use host_rt::mcu_call::McuCall;
-    use mcu_protocol::MessageKind;
-    use mcu_protocol::codec::{Cursor, Decode};
-    use mcu_protocol::messages::MotorStateResponse;
-    use runtime::stepping_state::MAX_AXES;
-
-    let configs = mcu_axis_configs
-        .lock()
-        .unwrap_or_else(|p| p.into_inner())
-        .clone();
-    if configs.is_empty() {
-        return Err("query_motor_positions: no axes configured".into());
-    }
-    let kin_tag = configs
-        .iter()
-        .find(|c| c.axes.contains(&0usize))
-        .map(|c| c.kinematics)
-        .unwrap_or(runtime::segment::KinematicTag::Cartesian as u8);
-
-    let mut motors: [Option<f64>; MAX_AXES] = [None; MAX_AXES];
-    let mut vmotors: [Option<f64>; MAX_AXES] = [None; MAX_AXES];
-
-    for cfg in &configs {
-        let q = {
-            let map = mcus.lock().unwrap_or_else(|p| p.into_inner());
-            let Some(conn) = map.get(&cfg.mcu_id) else {
-                continue;
-            };
-            if conn.ethercat_socket.is_some() {
-                match conn.endpoint_conn.as_ref() {
-                    Some(ep) => MotorQuery::EtherCat(Arc::clone(ep)),
-                    None => continue,
-                }
-            } else {
-                match conn.host_io.as_ref() {
-                    Some(io) => MotorQuery::Serial(Arc::clone(io)),
-                    None => continue,
-                }
-            }
-        };
-        let (kind, body) = match &q {
-            MotorQuery::Serial(io) => {
-                io.mcu_call(MessageKind::QueryMotorState, Vec::new(), timeout)
-            }
-            MotorQuery::EtherCat(ep) => {
-                ep.mcu_call(MessageKind::QueryMotorState, Vec::new(), timeout)
-            }
-        }
-        .map_err(|e| format!("query mcu {}: {e:?}", cfg.mcu_id))?;
-        if kind != MessageKind::MotorStateResponse {
-            return Err(format!(
-                "query mcu {}: unexpected kind {kind:?}",
-                cfg.mcu_id
-            ));
-        }
-        let mut c = Cursor::new(&body);
-        let resp = MotorStateResponse::decode_from(&mut c)
-            .map_err(|e| format!("query mcu {}: decode {e:?}", cfg.mcu_id))?;
-        place_motor_response(&resp, &cfg.axes, q.is_ethercat(), &mut motors, &mut vmotors);
-    }
-    crate::position_query::assemble_cartesian(&motors, &vmotors, kin_tag)
-}
-
-fn query_ethercat_runtime_caps(
-    conn: &McuSerialConn,
-    timeout: std::time::Duration,
-) -> Result<mcu_protocol::messages::RuntimeCapsResponse, RuntimeCapsError> {
-    use host_rt::mcu_call::McuCall;
-    use mcu_protocol::MessageKind;
-    let (kind, body) = conn
-        .mcu_call(MessageKind::QueryRuntimeCaps, Vec::new(), timeout)
-        .map_err(|e| RuntimeCapsError::Call(format!("{e:?}")))?;
-    if kind != MessageKind::RuntimeCapsResponse {
-        return Err(RuntimeCapsError::UnexpectedKind { got: kind });
-    }
-    decode_runtime_caps_body(&body)
-}
-
-#[derive(Debug)]
-enum EndpointClaimError {
-    DriveOffline { slave_idx: u8, fault_code: u16 },
-    DriveFault { slave_idx: u8, fault_code: u16 },
-    Protocol(String),
-}
-
-fn message_for_claim_error(label: &str, interface: &str, e: &EndpointClaimError) -> String {
-    match e {
-        EndpointClaimError::DriveOffline {
-            slave_idx,
-            fault_code,
-        } => match fault_code {
-            1 | 2 => format!(
-                "ethercat {label}: EtherCAT bus on {interface}: no slaves responding \
-                 (bringup rc=-{fault_code}) — check cable and drive power, then FIRMWARE_RESTART"
-            ),
-            10..=12 => format!(
-                "ethercat {label}: realtime endpoint could not acquire RT scheduling \
-                 (bringup rc=-{fault_code}) — grant CAP_SYS_NICE + CAP_IPC_LOCK to \
-                 klipper.service and isolate a CPU core, then FIRMWARE_RESTART"
-            ),
-            0 => format!(
-                "ethercat {label}: drive (slave {slave_idx}) offline \
-                 — check drive power, then FIRMWARE_RESTART"
-            ),
-            _ => format!(
-                "ethercat {label}: drive (slave {slave_idx}) offline \
-                 (bringup rc=-{fault_code}) — check drive power, then FIRMWARE_RESTART"
-            ),
-        },
-        EndpointClaimError::DriveFault {
-            slave_idx,
-            fault_code,
-        } => format!(
-            "ethercat {label}: drive (slave {slave_idx}) \
-             fault 0x{fault_code:04x} — check drive, then FIRMWARE_RESTART"
-        ),
-        EndpointClaimError::Protocol(s) => {
-            format!("ethercat {label}: endpoint protocol error — {s}")
-        }
-    }
-}
-
-fn push_drive_flags(args: &mut Vec<String>, d: &EthercatDrive) {
-    let (
-        _chain_index,
-        _axis,
-        counts_per_mm,
-        rotation_distance,
-        ferr,
-        max_torque,
-        velocity_ff,
-        ff_torque_clamp,
-        invert_direction,
-        dynamics_profile,
-    ) = d;
-    args.push("--counts-per-mm".into());
-    args.push(counts_per_mm.to_string());
-    args.push("--rotation-distance".into());
-    args.push(rotation_distance.to_string());
-    if let Some(ferr) = ferr {
-        args.push("--following-error-counts".into());
-        args.push(ferr.to_string());
-    }
-    if let Some(tq) = max_torque {
-        args.push("--max-torque-tenth-pct".into());
-        args.push(tq.to_string());
-    }
-    if *velocity_ff {
-        args.push("--velocity-ff".into());
-    }
-    if *invert_direction {
-        args.push("--invert".into());
-    }
-    args.push("--torque-clamp-pct".into());
-    args.push(ff_torque_clamp.to_string());
-    if let Some(profile) = dynamics_profile {
-        args.push("--slave-dynamics-profile".into());
-        args.push(profile.to_string());
-    }
-}
-
-fn endpoint_args(
-    interface: &str,
-    socket_path: &str,
-    dynamics_profile: Option<&str>,
-    events_dir: Option<&std::path::Path>,
-    drives: &[EthercatDrive],
-) -> Vec<String> {
-    let mut args = vec![
-        interface.to_string(),
-        "--socket".into(),
-        socket_path.to_string(),
-    ];
-    if let Some(p) = dynamics_profile {
-        args.push("--dynamics-profile".into());
-        args.push(p.to_string());
-    }
-    if let Some(dir) = events_dir {
-        args.push("--events-dir".into());
-        args.push(dir.to_string_lossy().into_owned());
-    }
-    if drives.len() == 1 {
-        push_drive_flags(&mut args, &drives[0]);
-    } else {
-        for d in drives {
-            let (chain_index, axis, ..) = d;
-            args.push("--slave".into());
-            args.push(chain_index.to_string());
-            args.push("--axis".into());
-            args.push(axis.to_string());
-            push_drive_flags(&mut args, d);
-        }
-    }
-    args
-}
-
-fn spawn_ethercat_endpoint(
-    binary: &str,
-    interface: &str,
-    socket_path: &str,
-    dynamics_profile: Option<&str>,
-    events_dir: Option<&std::path::Path>,
-    drives: &[EthercatDrive],
-) -> Result<std::process::Child, String> {
-    let args = endpoint_args(interface, socket_path, dynamics_profile, events_dir, drives);
-    std::process::Command::new(binary)
-        .args(&args)
-        .spawn()
-        .map_err(|e| format!("spawn {binary}: {e}"))
-}
-
-fn poll_socket_ready(
-    path: &str,
-    deadline: Instant,
-    child: &mut std::process::Child,
-) -> Result<(), String> {
-    loop {
-        if std::path::Path::new(path).exists() {
-            return Ok(());
-        }
-        if Instant::now() >= deadline {
-            return Err(format!("endpoint socket {path} did not appear within 15 s"));
-        }
-        match child.try_wait() {
-            Ok(Some(status)) => {
-                return Err(format!(
-                    "endpoint process exited before socket appeared \
-                     (exit status: {status})"
-                ));
-            }
-            Ok(None) => {}
-            Err(e) => {
-                return Err(format!("try_wait on endpoint process failed: {e}"));
-            }
-        }
-        std::thread::sleep(Duration::from_millis(20));
-    }
-}
-
-fn handshake_ethercat_endpoint(
-    socket_path: &str,
-    deadline: Instant,
-) -> Result<McuSerialConn, EndpointClaimError> {
-    use host_rt::mcu_call::McuCall;
-    use mcu_protocol::MessageKind;
-    use mcu_protocol::codec::{Cursor, Decode};
-    use mcu_protocol::messages::ClaimHandshakeReply;
-
-    let conn = loop {
-        match McuSerialConn::connect(socket_path) {
-            Ok(c) => break c,
-            Err(e)
-                if e.kind() == std::io::ErrorKind::ConnectionRefused
-                    || e.kind() == std::io::ErrorKind::NotFound =>
-            {
-                if Instant::now() >= deadline {
-                    return Err(EndpointClaimError::Protocol(format!(
-                        "connect to {socket_path}: timed out waiting for listener ({e})"
-                    )));
-                }
-                std::thread::sleep(Duration::from_millis(50));
-            }
-            Err(e) => {
-                return Err(EndpointClaimError::Protocol(format!(
-                    "connect to {socket_path}: {e}"
-                )));
-            }
-        }
-    };
-
-    let remaining = deadline.saturating_duration_since(Instant::now());
-    let (kind, body) = conn
-        .mcu_call(MessageKind::ClaimHandshake, Vec::new(), remaining)
-        .map_err(|e| EndpointClaimError::Protocol(format!("ClaimHandshake call: {e:?}")))?;
-
-    if kind != MessageKind::ClaimHandshakeReply {
-        return Err(EndpointClaimError::Protocol(format!(
-            "expected ClaimHandshakeReply (0x{:04x}), got 0x{:04x}",
-            MessageKind::ClaimHandshakeReply.as_u16(),
-            kind.as_u16(),
-        )));
-    }
-
-    let reply = ClaimHandshakeReply::decode_from(&mut Cursor::new(&body))
-        .map_err(|e| EndpointClaimError::Protocol(format!("decode ClaimHandshakeReply: {e:?}")))?;
-
-    for s in &reply.slave_statuses {
-        match s.state {
-            mcu_protocol::messages::SlaveState::Offline => {
-                return Err(EndpointClaimError::DriveOffline {
-                    slave_idx: s.slave_idx,
-                    fault_code: s.fault_code,
-                });
-            }
-            mcu_protocol::messages::SlaveState::Fault => {
-                return Err(EndpointClaimError::DriveFault {
-                    slave_idx: s.slave_idx,
-                    fault_code: s.fault_code,
-                });
-            }
-            mcu_protocol::messages::SlaveState::Ok => {}
-        }
-    }
-
-    Ok(conn)
-}
-
-#[derive(Debug, Clone)]
-struct EngineEvent {
-    kind: String,
-    mcu: u32,
-    notify_id: u64,
-    response_bytes: Vec<u8>,
-    sent_time: f64,
-    receive_time: f64,
-}
-
-impl EngineEvent {
-    fn to_pydict(&self, py: Python<'_>) -> PyResult<Py<PyDict>> {
-        let d = PyDict::new(py);
-        d.set_item("type", &self.kind)?;
-        d.set_item("mcu", self.mcu)?;
-        d.set_item("notify_id", self.notify_id)?;
-        d.set_item("data", pyo3::types::PyBytes::new(py, &self.response_bytes))?;
-        d.set_item("sent_time", self.sent_time)?;
-        d.set_item("receive_time", self.receive_time)?;
-        Ok(d.unbind())
-    }
-}
 
 fn router_err(e: host_rt::passthrough_queue::RouterError) -> PyErr {
     PyRuntimeError::new_err(e.to_string())
@@ -601,37 +78,6 @@ const STREAM_MAX_BUFFER_MOVES: usize = 512;
 /// five significant figures (the residual is non-monotonic integration noise,
 /// not a slower path).
 const STREAM_INTEGRATION_TOL: f64 = 1e-4;
-
-fn resolve_motion_caps(
-    caps: Option<mcu_protocol::messages::RuntimeCapsResponse>,
-    label: &str,
-    handle: u32,
-) -> Result<McuCaps, String> {
-    caps.map(McuCaps::from).ok_or_else(|| {
-        format!(
-            "no runtime caps for {label} MCU (handle={handle}) — cannot size piece rings; \
-             firmware not flashed or QueryRuntimeCaps failed at attach"
-        )
-    })
-}
-
-fn require_events_dir_for_mcu_transport(
-    mcu_transport: bool,
-    events_dir: Option<&std::path::Path>,
-    mcu_label: &str,
-) -> Result<(), String> {
-    if mcu_transport && events_dir.is_none() {
-        return Err(format!(
-            "attach_serial({mcu_label}): init_logging must be called before \
-             attach_serial for a kalico-native MCU — the dedicated \
-             mcu-*.jsonl writer cannot be installed without an events_dir. \
-             All McuLog events would be silently discarded to the general \
-             runtime_rx channel with no NDJSON output, which violates the \
-             observability spec (§4, Decision C). Call init_logging first."
-        ));
-    }
-    Ok(())
-}
 
 #[pyclass(name = "MotionEngine")]
 #[allow(missing_debug_implementations)]
@@ -684,56 +130,6 @@ pub struct PyMotionEngine {
     latched_endpoint_death: Arc<Mutex<HashMap<u32, String>>>,
     remote_triggers: Mutex<HashMap<u8, (u32, host_rt::host_io::InterceptorId)>>,
     shut_down: AtomicBool,
-}
-
-pub(crate) fn axis_ring_depth(total_pieces: u32, num_axes: u32) -> u32 {
-    (total_pieces / num_axes.max(1)).max(1)
-}
-
-pub(crate) fn drip_cohort_participants(configs: &[McuAxisConfig]) -> Vec<crate::types::AxisKey> {
-    configs
-        .iter()
-        .flat_map(|cfg| {
-            cfg.axes.iter().map(move |&a| crate::types::AxisKey {
-                mcu_id: cfg.mcu_id,
-                axis: a as u8,
-            })
-        })
-        .collect()
-}
-
-pub(crate) fn ring_depth_for_axis_inner(
-    configs: &[crate::mcu_config::McuAxisConfig],
-    mcu_handle: u32,
-    axis: u8,
-) -> Result<u16, String> {
-    let cfg = configs
-        .iter()
-        .find(|c| c.mcu_id == mcu_handle)
-        .ok_or_else(|| {
-            format!(
-                "ring_depth_for_axis: unknown mcu_handle {mcu_handle} \
-                 (init_planner not yet called?)"
-            )
-        })?;
-    let axis_usize = usize::from(axis);
-    if !cfg.axes.contains(&axis_usize) {
-        return Err(format!(
-            "ring_depth_for_axis: axis {axis} is not configured on mcu_handle \
-             {mcu_handle} (configured axes: {:?})",
-            cfg.axes
-        ));
-    }
-    let depth = axis_ring_depth(cfg.caps.total_pieces() as u32, cfg.axes.len() as u32);
-    if depth > u32::from(u16::MAX) {
-        return Err(format!(
-            "ring depth {depth} exceeds u16::MAX (65535) for mcu {mcu_handle} axis {axis}; \
-             a >65535-piece ring would need >2 MB of SRAM and is impossible here — \
-             check total_piece_memory configuration"
-        ));
-    }
-    #[allow(clippy::cast_possible_truncation)]
-    Ok(depth as u16)
 }
 
 #[pymethods]
@@ -1019,20 +415,7 @@ impl PyMotionEngine {
         following_error_counts: u32,
         max_torque_tenth_pct: u16,
     ) -> PyResult<()> {
-        let conn = {
-            let mcus = self.mcus.lock().unwrap_or_else(|p| p.into_inner());
-            let mc = mcus.get(&mcu_handle).ok_or_else(|| {
-                PyRuntimeError::new_err(format!(
-                    "set_drive_limits: unknown mcu_handle {mcu_handle}"
-                ))
-            })?;
-            mc.endpoint_conn.clone().ok_or_else(|| {
-                PyRuntimeError::new_err(format!(
-                    "set_drive_limits: mcu {mcu_handle} ({}) is not an EtherCAT endpoint",
-                    mc.label
-                ))
-            })?
-        };
+        let conn = self.ethercat_conn(mcu_handle, "set_drive_limits")?;
         tracing::info!(
             subsystem = "engine",
             event = "servo_drive_limits",
@@ -1057,20 +440,7 @@ impl PyMotionEngine {
     }
 
     fn restore_drive_limits(&self, mcu_handle: u32, slot: u8) -> PyResult<()> {
-        let conn = {
-            let mcus = self.mcus.lock().unwrap_or_else(|p| p.into_inner());
-            let mc = mcus.get(&mcu_handle).ok_or_else(|| {
-                PyRuntimeError::new_err(format!(
-                    "restore_drive_limits: unknown mcu_handle {mcu_handle}"
-                ))
-            })?;
-            mc.endpoint_conn.clone().ok_or_else(|| {
-                PyRuntimeError::new_err(format!(
-                    "restore_drive_limits: mcu {mcu_handle} ({}) is not an EtherCAT endpoint",
-                    mc.label
-                ))
-            })?
-        };
+        let conn = self.ethercat_conn(mcu_handle, "restore_drive_limits")?;
         tracing::info!(
             subsystem = "engine",
             event = "servo_drive_limits",
@@ -1749,34 +1119,33 @@ impl PyMotionEngine {
                         "attach_serial: reuse — criticality set"
                     );
 
-                    let mut mcus = self.mcus.lock().unwrap_or_else(|p| p.into_inner());
-                    let conn = mcus.get_mut(&mcu_handle).ok_or_else(|| {
-                        PyRuntimeError::new_err(format!(
-                            "attach_serial: unknown mcu_handle {mcu_handle}"
-                        ))
-                    })?;
-                    conn.runtime_rx_priority = Some(rx_priority);
-                    conn.runtime_rx_bulk = Some(rx_bulk);
-                    conn.runtime_caps = runtime_caps;
-                    conn.identify_caps = identify_caps;
-                    conn.mcu_transport_supported = mcu_transport_supported;
+                    self.with_mcu(
+                        mcu_handle,
+                        |h| format!("attach_serial: unknown mcu_handle {h}"),
+                        |conn| {
+                            conn.runtime_rx_priority = Some(rx_priority);
+                            conn.runtime_rx_bulk = Some(rx_bulk);
+                            conn.runtime_caps = runtime_caps;
+                            conn.identify_caps = identify_caps;
+                            conn.mcu_transport_supported = mcu_transport_supported;
+                            Ok(())
+                        },
+                    )?;
                     return Ok(());
                 }
             }
         }
 
-        let mcu_label: String = {
-            let mut mcus = self.mcus.lock().unwrap_or_else(|p| p.into_inner());
-            let conn = mcus.get_mut(&mcu_handle).ok_or_else(|| {
-                PyRuntimeError::new_err(format!(
-                    "attach_serial: unknown mcu_handle {mcu_handle} (claim_mcu not called)"
-                ))
-            })?;
-            conn.runtime_rx_priority = None;
-            conn.runtime_rx_bulk = None;
-            conn.host_io = None;
-            conn.label.clone()
-        };
+        let mcu_label: String = self.with_mcu(
+            mcu_handle,
+            |h| format!("attach_serial: unknown mcu_handle {h} (claim_mcu not called)"),
+            |conn| {
+                conn.runtime_rx_priority = None;
+                conn.runtime_rx_bulk = None;
+                conn.host_io = None;
+                Ok(conn.label.clone())
+            },
+        )?;
 
         let is_pipe = baud == 0
             || serial_path.starts_with("/tmp/")
@@ -1951,27 +1320,27 @@ impl PyMotionEngine {
             }
         }
 
-        let mut mcus = self.mcus.lock().unwrap_or_else(|p| p.into_inner());
-        let conn = mcus.get_mut(&mcu_handle).ok_or_else(|| {
-            PyRuntimeError::new_err(format!("attach_serial: unknown mcu_handle {mcu_handle}"))
-        })?;
-        conn.host_io = Some(host_io_arc);
-        conn.runtime_rx_priority = Some(rx_priority);
-        conn.runtime_rx_bulk = Some(rx_bulk);
-        conn.runtime_caps = runtime_caps;
-        conn.identify_caps = identify_caps;
-        conn.mcu_transport_supported = mcu_transport_supported;
-        Ok(())
+        self.with_mcu(
+            mcu_handle,
+            |h| format!("attach_serial: unknown mcu_handle {h}"),
+            |conn| {
+                conn.host_io = Some(host_io_arc);
+                conn.runtime_rx_priority = Some(rx_priority);
+                conn.runtime_rx_bulk = Some(rx_bulk);
+                conn.runtime_caps = runtime_caps;
+                conn.identify_caps = identify_caps;
+                conn.mcu_transport_supported = mcu_transport_supported;
+                Ok(())
+            },
+        )
     }
 
     fn get_mcu_capabilities(&self, mcu_handle: u32) -> PyResult<u64> {
-        let mcus = self.mcus.lock().unwrap_or_else(|p| p.into_inner());
-        let conn = mcus.get(&mcu_handle).ok_or_else(|| {
-            PyRuntimeError::new_err(format!(
-                "get_mcu_capabilities: unknown mcu_handle {mcu_handle}"
-            ))
-        })?;
-        Ok(conn.identify_caps)
+        self.with_mcu(
+            mcu_handle,
+            |h| format!("get_mcu_capabilities: unknown mcu_handle {h}"),
+            |conn| Ok(conn.identify_caps),
+        )
     }
 
     fn ring_depth_for_axis(&self, mcu_handle: u32, axis: u8) -> PyResult<u16> {
@@ -2130,22 +1499,7 @@ impl PyMotionEngine {
     }
 
     fn get_identify_data(&self, mcu_handle: u32) -> PyResult<Vec<u8>> {
-        let io = {
-            let mcus = self.mcus.lock().unwrap_or_else(|p| p.into_inner());
-            let conn = mcus.get(&mcu_handle).ok_or_else(|| {
-                PyRuntimeError::new_err(format!(
-                    "get_identify_data: unknown mcu_handle {mcu_handle}"
-                ))
-            })?;
-            conn.host_io
-                .as_ref()
-                .ok_or_else(|| {
-                    PyRuntimeError::new_err(
-                        "get_identify_data: attach_serial has not been called for this MCU",
-                    )
-                })?
-                .clone()
-        };
+        let io = self.host_io_for_mcu("get_identify_data", mcu_handle)?;
         Ok(io.raw_identify_bytes().to_vec())
     }
 
@@ -2160,20 +1514,7 @@ impl PyMotionEngine {
     ) -> PyResult<Py<PyDict>> {
         use std::time::Duration;
 
-        let io = {
-            let mcus = self.mcus.lock().unwrap_or_else(|p| p.into_inner());
-            let conn = mcus.get(&mcu_handle).ok_or_else(|| {
-                PyRuntimeError::new_err(format!("engine_call: unknown mcu_handle {mcu_handle}"))
-            })?;
-            conn.host_io
-                .as_ref()
-                .ok_or_else(|| {
-                    PyRuntimeError::new_err(
-                        "engine_call: attach_serial has not been called for this MCU",
-                    )
-                })?
-                .clone()
-        };
+        let io = self.host_io_for_mcu("engine_call", mcu_handle)?;
 
         let msg_owned = msg.to_owned();
         let response_owned = response.to_owned();
@@ -2300,20 +1641,7 @@ impl PyMotionEngine {
     }
 
     fn engine_get_clock_async(&self, mcu_handle: u32) -> PyResult<()> {
-        let io =
-            {
-                let mcus = self.mcus.lock().unwrap_or_else(|p| p.into_inner());
-                let conn = mcus.get(&mcu_handle).ok_or_else(|| {
-                    pyo3::exceptions::PyRuntimeError::new_err(format!(
-                        "engine_get_clock_async: unknown mcu_handle {mcu_handle}"
-                    ))
-                })?;
-                conn.host_io.as_ref().ok_or_else(|| {
-                pyo3::exceptions::PyRuntimeError::new_err(
-                    "engine_get_clock_async: attach_serial has not been called for this MCU",
-                )
-            })?.clone()
-            };
+        let io = self.host_io_for_mcu("engine_get_clock_async", mcu_handle)?;
 
         io.get_clock_async().map_err(|e| {
             pyo3::exceptions::PyRuntimeError::new_err(format!("engine_get_clock_async: {e}"))
@@ -2322,38 +1650,13 @@ impl PyMotionEngine {
 
     #[pyo3(signature = (mcu_handle, msg))]
     fn engine_send(&self, mcu_handle: u32, msg: &str) -> PyResult<()> {
-        let io = {
-            let mcus = self.mcus.lock().unwrap_or_else(|p| p.into_inner());
-            let conn = mcus.get(&mcu_handle).ok_or_else(|| {
-                PyRuntimeError::new_err(format!("engine_send: unknown mcu_handle {mcu_handle}"))
-            })?;
-            conn.host_io
-                .as_ref()
-                .ok_or_else(|| {
-                    PyRuntimeError::new_err(
-                        "engine_send: attach_serial has not been called for this MCU",
-                    )
-                })?
-                .clone()
-        };
+        let io = self.host_io_for_mcu("engine_send", mcu_handle)?;
         io.send_fire_and_forget(msg)
             .map_err(|e| PyRuntimeError::new_err(format!("engine_send: {e}")))
     }
 
     fn engine_mark_expected_disconnect(&self, mcu_handle: u32) -> PyResult<()> {
-        let io = {
-            let mcus = self.mcus.lock().unwrap_or_else(|p| p.into_inner());
-            let conn = mcus.get(&mcu_handle).ok_or_else(|| {
-                PyRuntimeError::new_err(format!(
-                    "engine_mark_expected_disconnect: unknown mcu_handle {mcu_handle}"
-                ))
-            })?;
-            conn.host_io.as_ref().ok_or_else(|| {
-                PyRuntimeError::new_err(
-                    "engine_mark_expected_disconnect: attach_serial has not been called for this MCU",
-                )
-            })?.clone()
-        };
+        let io = self.host_io_for_mcu("engine_mark_expected_disconnect", mcu_handle)?;
         io.mark_expected_disconnect()
             .map_err(|e| PyRuntimeError::new_err(format!("engine_mark_expected_disconnect: {e}")))
     }
@@ -4129,11 +3432,6 @@ impl Drop for PyMotionEngine {
     }
 }
 
-struct FlushWait {
-    rx: crossbeam_channel::Receiver<Option<std::time::Instant>>,
-    deadline: Option<std::time::Instant>,
-}
-
 #[derive(Clone)]
 pub(crate) struct TripDeps {
     homing_run: Arc<Mutex<Option<HomingRun>>>,
@@ -4179,6 +3477,19 @@ impl PyMotionEngine {
                 .map_err(planner_err),
             None => Ok(()),
         }
+    }
+
+    fn with_mcu<R>(
+        &self,
+        handle: u32,
+        unknown_mcu_err: impl FnOnce(u32) -> String,
+        f: impl FnOnce(&mut McuConnection) -> PyResult<R>,
+    ) -> PyResult<R> {
+        let mut mcus = self.mcus.lock().unwrap_or_else(|p| p.into_inner());
+        let mc = mcus
+            .get_mut(&handle)
+            .ok_or_else(|| PyRuntimeError::new_err(unknown_mcu_err(handle)))?;
+        f(mc)
     }
 
     fn ethercat_conn(&self, mcu_handle: u32, what: &str) -> PyResult<Arc<McuSerialConn>> {
